projects/final_project/Final-Project-main/server/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Faker==28.4.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flask_jwt_extended==4.7.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
oauthlib==3.2.2
packaging==24.1
playwright==1.46.0
pluggy==1.5.0
psycopg==3.2.1
pyee==11.1.0
pyjokes==0.8.3
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
pytest-timeout==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
WTForms==3.1.2
